consolidate sol! struct and rm sp1-contract-call for now
