+++ canoe/bindings/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "canoe-bindings"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+alloy-sol-types = { workspace = true, features = ["json"]}
