--- canoe/contracts/eigenda-bindings/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "canoe-contracts-eigenda-bindings"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-alloy-sol-types.workspace = true
