--- canoe/contracts/eigenda-bindings/src/lib.rs
@@ -1,6 +1,0 @@
-use alloy_sol_types::sol;
-
-sol! {
-    IEigenDACertVerifier,
-    "IEigenDACertVerifier.json"
-}
