--- canoe/sp1-contract-call/host/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-version = "0.1.0"
-name = "canoe-sp1-contract-call-host"
-edition = "2021"
-
-[dependencies]
-# workspace
-sp1-cc-host-executor.workspace = true
-sp1-cc-client-executor.workspace = true
-
-# misc:
-
-# sp1
-sp1-sdk.workspace = true
-
-
-[build-dependencies]
-sp1-build.workspace = true
-
-
-[features]
-default = []
-cuda = ["sp1-sdk/cuda"]
