+++ canoe/steel/apps/Cargo.toml
@@ -11,13 +11,13 @@ url = {workspace = true }
 
 anyhow = { workspace = true }
 async-trait = { workspace = true }
-clap = { workspace = true, features = ["derive", "env"] }
 canoe-steel-methods = { workspace = true }
 risc0-steel = { workspace = true, features = ["host"] }
 risc0-zkvm = { workspace = true, features = ["client", "bonsai"] }
 tokio = { workspace = true }
 eigenda-v2-struct = { workspace = true }
 canoe-provider.workspace = true
+canoe-bindings.workspace = true
 
 [features]
 beacon = []
