+++ canoe/steel/methods/Cargo.toml
@@ -11,7 +11,6 @@ hex = { workspace = true }
 risc0-build = { workspace = true }
 risc0-build-ethereum = { workspace = true }
 risc0-zkp = { workspace = true }
-eigenda-v2-struct = { workspace = true }
 
 [dev-dependencies]
 alloy-primitives = { workspace = true }
