+++ canoe/steel/methods/guest/src/bin/dacert_v2_verifier.rs
@@ -16,97 +16,17 @@
 #![no_main]
 
 use alloy_primitives::Address;
-use alloy_sol_types::{sol, SolValue};
+use alloy_sol_types::SolValue;
 use risc0_steel::{
     ethereum::{EthEvmInput, ETH_HOLESKY_CHAIN_SPEC},
     Contract,
 };
 use risc0_zkvm::guest::env;
+use canoe_bindings::{
+    Journal, BatchHeaderV2, BlobInclusionInfo, NonSignerStakesAndSignature, IEigenDACertMockVerifier,
+};
 
 risc0_zkvm::guest::entry!(main);
-
-/// Specify the function to call using the [`sol!`] macro.
-/// This parses the Solidity syntax to generate a struct that implements the `SolCall` trait.
-
-sol! {
-    struct BatchHeaderV2 {
-        bytes32 batchRoot;
-        uint32 referenceBlockNumber;
-    }
-
-    struct BlobInclusionInfo {
-        BlobCertificate blobCertificate;
-        uint32 blobIndex;
-        bytes inclusionProof;
-    }
-
-    struct BlobCertificate {
-        BlobHeaderV2 blobHeader;
-        bytes signature;
-        uint32[] relayKeys;
-    }
-
-    struct BlobHeaderV2 {
-        uint16 version;
-        bytes quorumNumbers;
-        BlobCommitment commitment;
-        bytes32 paymentHeaderHash;        
-    }
-
-    struct G1Point {
-        uint256 X;
-        uint256 Y;
-    }
-
-    // Encoding of field elements is: X[1] * i + X[0]
-    struct G2Point {
-        uint256[2] X;
-        uint256[2] Y;
-    }
-
-    struct BlobCommitment {
-        G1Point commitment;
-        G2Point lengthCommitment;
-        G2Point lengthProof;
-        uint32 length;
-    }
-
-    struct NonSignerStakesAndSignature {
-        uint32[] nonSignerQuorumBitmapIndices; 
-        G1Point[] nonSignerPubkeys; 
-        G1Point[] quorumApks; 
-        G2Point apkG2; 
-        G1Point sigma; 
-        uint32[] quorumApkIndices; 
-        uint32[] totalStakeIndices; 
-        uint32[][] nonSignerStakeIndices; 
-    }
-
-    interface IEigenDACertVerifier { 
-        function alwaysReturnsTrue(
-            BatchHeaderV2 calldata batchHeader,
-            BlobInclusionInfo calldata blobInclusionInfo,
-            NonSignerStakesAndSignature calldata nonSignerStakesAndSignature,
-            bytes signedQuorumNumbers
-        ) external view returns (bool);
-
-        function verifyDACertV2ForZKProof(
-            BatchHeaderV2 calldata batchHeader,
-            BlobInclusionInfo calldata blobInclusionInfo,
-            NonSignerStakesAndSignature calldata nonSignerStakesAndSignature,
-            bytes signedQuorumNumbers
-        ) external view returns (bool);
-    }
-}
-
-/// ABI encodable journal data.
-sol! {
-    struct Journal {
-        address contract;
-        bytes input;        
-        // add chain spec    
-    }
-}
 
 fn main() {
     // Read the input from the guest environment.
@@ -128,7 +48,7 @@ fn main() {
     let env = input.into_env().with_chain_spec(&ETH_HOLESKY_CHAIN_SPEC);
 
     // Execute the view call; it returns the result in the type generated by the `sol!` macro.        
-    let call = IEigenDACertVerifier::alwaysReturnsTrueCall {        
+    let call = IEigenDACertMockVerifier::alwaysReturnsTrueCall {        
         batchHeader: batch_header,
         blobInclusionInfo: blob_inclusion_info.clone(),
         nonSignerStakesAndSignature: non_signer_stakes_and_signature,
